Error solving 2

+++ src/main/resources/db/hsqldb/initDB.sql
@@ -85,7 +85,7 @@ CREATE TABLE offers (
   id            INTEGER IDENTITY PRIMARY KEY,
   title	     	VARCHAR(255) NOT NULL,
   description	VARCHAR(255) NOT NULL,
-  discount		DOUBLE(5,2) NOT NULL,
+  discount		DOUBLE NOT NULL,
   expiration	DATE NOT NULL,
   valid			BOOLEAN 
 );
